# pay-guard-backend/docs/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
    "/api/v1/auth/login": {
      "post": {
        "operationId": "Auth_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Authentication"
        ]
      }
    },
    "/api/v1/auth/refresh": {
      "post": {
        "operationId": "Auth_refresh",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "tags": [
          "Authentication"
        ]
      }
    },
    "/api/v1/auth/logout": {
      "post": {
        "operationId": "Auth_logout",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LogoutDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Authentication"
        ]
      }
    },
    "/api/v1/auth/me": {
      "get": {
        "operationId": "Auth_me",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Authentication"
        ]
      }
    },
...
    "/api/v1/businesses/{businessId}/branches": {
      "post": {
        "operationId": "Branches_create",
        "parameters": [
          {
            "name": "businessId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateBranchDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Branches"
        ]
      },
      "get": {
        "operationId": "Branches_list",
        "parameters": [
          {
            "name": "businessId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Branches"
        ]
      }
    },
    "/api/v1/businesses/{businessId}/branches/{branchId}": {
      "patch": {
        "operationId": "Branches_update",
        "parameters": [
          {
            "name": "businessId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "branchId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateBranchDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Branches"
        ]
      }
    },
...
    "/api/v1/businesses/{businessId}/ledger": {
      "get": {
        "operationId": "Ledger_list",
        "parameters": [
          {
            "name": "businessId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "branchId",
            "required": false,
            "in": "query",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "settlementAccountId",
            "required": false,
            "in": "query",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "entryType",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "OPENING_BALANCE",
                "VERIFIED_DEPOSIT",
                "MANUAL_DEPOSIT",
                "WITHDRAWAL",
                "POSITIVE_CORRECTION",
                "NEGATIVE_CORRECTION",
                "REVERSAL"
              ]
            }
          },
          {
            "name": "dateFrom",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "dateTo",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 50,
              "allOf": [
                {
                  "$ref": "#/components/schemas/Object"
                }
              ]
            }
          },
          {
            "name": "offset",
            "required": true,
            "in": "query",
            "schema": {
              "minimum": 0,
              "maximum": 10000,
              "default": 0,
              "allOf": [
                {
                  "$ref": "#/components/schemas/Object"
                }
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Ledger"
        ]
      }
    },
    "/api/v1/businesses/{businessId}/ledger/accounts/{accountId}/projected-balance": {
      "get": {
        "operationId": "Ledger_projectedBalance",
        "parameters": [
          {
            "name": "businessId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "accountId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "direction",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "CREDIT",
                "DEBIT"
              ]
            }
          },
          {
            "name": "amount",
            "required": true,
            "in": "query",
            "schema": {
              "pattern": "^\\d{1,16}\\.\\d{2}$",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Ledger"
        ]
      }
    },
    "/api/v1/businesses/{businessId}/ledger/{entryId}": {
      "get": {
        "operationId": "Ledger_require",
        "parameters": [
          {
            "name": "businessId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "entryId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Ledger"
        ]
      }
    },
...
    "/api/v1/businesses/{businessId}/branches/{branchId}/credits": {
      "get": {
        "operationId": "Credit_wallet",
        "parameters": [
          {
            "name": "businessId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "branchId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Branch Credits"
        ]
      }
    },
    "/api/v1/businesses/{businessId}/branches/{branchId}/credits/history": {
      "get": {
        "operationId": "Credit_history",
        "parameters": [
          {
            "name": "businessId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "branchId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "eventType",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "SUBSCRIPTION_CREDIT_GRANT",
                "VERIFICATION_DEDUCTION",
                "SUBSCRIPTION_VERIFICATION_DEFERRED",
                "DEFERRED_DEDUCTION_SETTLED",
                "CREDIT_EXPIRY",
                "ADMIN_ADJUSTMENT"
              ]
            }
          },
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 50,
              "allOf": [
                {
                  "$ref": "#/components/schemas/Object"
                }
              ]
            }
          },
          {
            "name": "offset",
            "required": true,
            "in": "query",
            "schema": {
              "minimum": 0,
              "maximum": 10000,
              "default": 0,
              "allOf": [
                {
                  "$ref": "#/components/schemas/Object"
                }
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Branch Credits"
        ]
      }
    },
...
  },
  "info": {
    "title": "PayGuard API",
    "description": "Canonical API contract for the PayGuard web portal, Android app, workers, and third-party integrations.",
    "version": "2.0.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:4000",
      "description": "Local development"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Access token returned by POST /api/v1/auth/login."
      }
    },
    "schemas": {
      "LoginContextDto": {
        "type": "object",
        "properties": {
          "membershipId": {
            "type": "string",
            "format": "uuid"
          },
          "membershipRoleId": {
            "type": "string",
            "format": "uuid"
          },
          "workAssignmentId": {
            "type": "string",
            "format": "uuid"
          }
        },
        "required": [
          "membershipId",
          "membershipRoleId"
        ]
      },
      "LoginDto": {
        "type": "object",
        "properties": {
          "identity": {
            "type": "string",
            "minLength": 3,
            "maxLength": 254
          },
          "password": {
            "type": "string",
            "minLength": 8
          },
          "deviceIdentifier": {
            "type": "string",
            "minLength": 16,
            "maxLength": 512
          },
          "devicePlatform": {
            "type": "string",
            "enum": [
              "android",
              "ios",
              "web"
            ]
          },
          "context": {
            "$ref": "#/components/schemas/LoginContextDto"
          }
        },
        "required": [
          "identity",
          "password"
        ]
      },
      "RefreshDto": {
        "type": "object",
        "properties": {
          "refreshToken": {
            "type": "string",
            "minLength": 32
          }
        },
        "required": [
          "refreshToken"
        ]
      },
      "LogoutDto": {
        "type": "object",
        "properties": {
          "refreshToken": {
            "type": "string",
            "minLength": 32
          }
        },
        "required": [
          "refreshToken"
        ]
      },
...
      "CreateBranchDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 2
          },
          "code": {
            "type": "string"
          },
          "address": {
            "type": "string"
          },
          "timezone": {
            "type": "string"
          },
          "city": {
            "type": "string"
          },
          "subCity": {
            "type": "string"
          },
          "woreda": {
            "type": "string"
          },
          "locationDetails": {
            "type": "string"
          },
          "settlementMode": {
            "type": "string",
            "enum": [
              "MAIN_BUSINESS_ALL",
              "BRANCH_SPECIFIC"
            ]
          }
        },
        "required": [
          "name"
        ]
      },
      "UpdateBranchDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 2
          },
          "address": {
            "type": "string"
          },
          "verificationTimeToleranceMinutes": {
            "type": "number",
            "minimum": 0,
            "maximum": 1440
          },
          "city": {
            "type": "string"
          },
          "subCity": {
            "type": "string"
          },
          "woreda": {
            "type": "string"
          },
          "locationDetails": {
            "type": "string"
          },
          "settlementMode": {
            "type": "string",
            "enum": [
              "MAIN_BUSINESS_ALL",
              "BRANCH_SPECIFIC"
            ]
          }
        }
      },
...
      "Object": {
        "type": "object",
        "properties": {}
      },
...
    }
  }
}